move Manifest inside module/
