+++ src/module/manifest.rs
