+++ src/sub/append_mod.rs
@@ -12,7 +12,7 @@ use crate::components::{Components, Component, FullComponent};
 use crate::language::{select_language_pref, LanguageSelection};
 use crate::list_components::list_components;
 use crate::lowercase::LwcString;
-use crate::manifest::Manifest;
+use crate::module::manifest::Manifest;
 use crate::module::module::Module;
 use crate::module::weidu_mod::WeiduMod;
 use crate::pathext::append_extension;
